extracted DB backup methods
